Autofocus / tabbing to change focus
This implements a 'focus chain', which is a list of widgets that
can exchange focus amongst themselves.

Widgets that want to participate in this mechanism can do so
by calling LifeCycleCtx::register_for_focus while handling the
LifeCycle::Register event.

At runtime, the currently focused widget can call methods on
EventCtx to give focus to the next or previous widget.